requirements.txt
pip
viam-sdk
boto3
pillow